Move `locate` to `current`
